refactor: refactor request creation into factory class
