+++ src/main/java/io/odpf/depot/http/HttpSinkFactory.java
@@ -2,18 +2,14 @@ package io.odpf.depot.http;
 
 import com.timgroup.statsd.NoOpStatsDClient;
 import io.odpf.depot.OdpfSink;
-import io.odpf.depot.common.Tuple;
 import io.odpf.depot.common.client.HttpClientUtils;
 import io.odpf.depot.config.HttpClientConfig;
 import io.odpf.depot.config.HttpSinkConfig;
 import io.odpf.depot.http.client.HttpSinkClient;
-import io.odpf.depot.http.parser.HttpRequestParser;
-import io.odpf.depot.message.OdpfMessageParser;
-import io.odpf.depot.message.OdpfMessageParserFactory;
-import io.odpf.depot.message.SinkConnectorSchemaMessageMode;
+import io.odpf.depot.http.request.Request;
+import io.odpf.depot.http.request.RequestFactory;
 import io.odpf.depot.metrics.Instrumentation;
 import io.odpf.depot.metrics.StatsDReporter;
-import io.odpf.depot.utils.MessageConfigUtils;
 import org.apache.http.impl.client.CloseableHttpClient;
 
 public class HttpSinkFactory {
@@ -23,7 +19,7 @@ public class HttpSinkFactory {
     private final StatsDReporter statsDReporter;
 
     private HttpSinkClient httpSinkClient;
-    private HttpRequestParser requestParser;
+    private Request request;
 
     public HttpSinkFactory(HttpSinkConfig sinkConfig, HttpClientConfig clientConfig, StatsDReporter statsDReporter) {
         this.sinkConfig = sinkConfig;
@@ -41,9 +37,7 @@ public class HttpSinkFactory {
         try {
             CloseableHttpClient closeableHttpClient = HttpClientUtils.newHttpClient(clientConfig, statsDReporter);
             httpSinkClient = new HttpSinkClient(closeableHttpClient, new Instrumentation(statsDReporter, HttpSinkClient.class));
-            OdpfMessageParser messageParser = OdpfMessageParserFactory.getParser(sinkConfig, statsDReporter);
-            Tuple<SinkConnectorSchemaMessageMode, String> modeAndSchema = MessageConfigUtils.getModeAndSchema(sinkConfig);
-            requestParser = new HttpRequestParser(messageParser, modeAndSchema, sinkConfig);
+            request = RequestFactory.create(sinkConfig);
 
         } catch (Exception e) {
             throw new IllegalArgumentException("Exception occurred while creating Http sink", e);
@@ -53,7 +47,7 @@ public class HttpSinkFactory {
     public OdpfSink create() {
         return new HttpSink(
                 httpSinkClient,
-                requestParser,
+                request,
                 new Instrumentation(statsDReporter, HttpSink.class)
         );
     }
