+++ src/test/java/io/odpf/depot/http/HttpSinkTest.java
@@ -5,8 +5,8 @@ import io.odpf.depot.error.ErrorInfo;
 import io.odpf.depot.error.ErrorType;
 import io.odpf.depot.exception.ConfigurationException;
 import io.odpf.depot.http.client.HttpSinkClient;
-import io.odpf.depot.http.parser.HttpRequestParser;
 import io.odpf.depot.http.record.HttpRequestRecord;
+import io.odpf.depot.http.request.Request;
 import io.odpf.depot.http.response.HttpSinkResponse;
 import io.odpf.depot.message.OdpfMessage;
 import io.odpf.depot.metrics.Instrumentation;
@@ -32,10 +32,10 @@ import static org.mockito.Mockito.when;
 public class HttpSinkTest {
 
     @Mock
-    private HttpEntityEnclosingRequestBase request;
+    private HttpEntityEnclosingRequestBase httpRequest;
 
     @Mock
-    private HttpRequestParser requestParser;
+    private Request request;
 
     @Mock
     private HttpSinkClient httpSinkClient;
@@ -56,20 +56,20 @@ public class HttpSinkTest {
     public void shouldPushToSink() throws IOException {
         List<OdpfMessage> messages = new ArrayList<>();
         List<HttpRequestRecord> records = new ArrayList<>();
-        records.add(new HttpRequestRecord(request, 0L, null, true));
-        records.add(new HttpRequestRecord(request, 1L, null, true));
-        records.add(new HttpRequestRecord(request, 2L, null, true));
-        records.add(new HttpRequestRecord(request, 3L, null, true));
-        records.add(new HttpRequestRecord(request, 4L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 0L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 1L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 2L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 3L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 4L, null, true));
         List<HttpSinkResponse> responses = new ArrayList<>();
         responses.add(Mockito.mock(HttpSinkResponse.class));
         responses.add(Mockito.mock(HttpSinkResponse.class));
         responses.add(Mockito.mock(HttpSinkResponse.class));
         responses.add(Mockito.mock(HttpSinkResponse.class));
         responses.add(Mockito.mock(HttpSinkResponse.class));
-        when(requestParser.convert(messages)).thenReturn(records);
+        when(request.createRecords(messages)).thenReturn(records);
         when(httpSinkClient.send(records)).thenReturn(responses);
-        HttpSink httpSink = new HttpSink(httpSinkClient, requestParser, instrumentation);
+        HttpSink httpSink = new HttpSink(httpSinkClient, request, instrumentation);
         OdpfSinkResponse odpfSinkResponse = httpSink.pushToSink(messages);
         Assert.assertFalse(odpfSinkResponse.hasErrors());
     }
@@ -79,18 +79,18 @@ public class HttpSinkTest {
         List<OdpfMessage> messages = new ArrayList<>();
         List<HttpRequestRecord> records = new ArrayList<>();
         records.add(new HttpRequestRecord(null, 0L, new ErrorInfo(new ConfigurationException(""), ErrorType.DEFAULT_ERROR), false));
-        records.add(new HttpRequestRecord(request, 1L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 1L, null, true));
         records.add(new HttpRequestRecord(null, 2L, new ErrorInfo(new ConfigurationException(""), ErrorType.DEFAULT_ERROR), false));
-        records.add(new HttpRequestRecord(request, 3L, null, true));
-        records.add(new HttpRequestRecord(request, 4L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 3L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 4L, null, true));
         List<HttpSinkResponse> responses = new ArrayList<>();
         responses.add(Mockito.mock(HttpSinkResponse.class));
         responses.add(Mockito.mock(HttpSinkResponse.class));
         responses.add(Mockito.mock(HttpSinkResponse.class));
-        when(requestParser.convert(messages)).thenReturn(records);
+        when(request.createRecords(messages)).thenReturn(records);
         List<HttpRequestRecord> validRecords = records.stream().filter(HttpRequestRecord::isValid).collect(Collectors.toList());
         when(httpSinkClient.send(validRecords)).thenReturn(responses);
-        HttpSink httpSink = new HttpSink(httpSinkClient, requestParser, instrumentation);
+        HttpSink httpSink = new HttpSink(httpSinkClient, request, instrumentation);
         OdpfSinkResponse odpfSinkResponse = httpSink.pushToSink(messages);
         Assert.assertTrue(odpfSinkResponse.hasErrors());
         Assert.assertEquals(2, odpfSinkResponse.getErrors().size());
@@ -102,13 +102,13 @@ public class HttpSinkTest {
     public void shouldReportErrors() throws IOException {
         List<OdpfMessage> messages = new ArrayList<>();
         List<HttpRequestRecord> records = new ArrayList<>();
-        records.add(new HttpRequestRecord(request, 0L, null, true));
-        records.add(new HttpRequestRecord(request, 1L, null, true));
-        records.add(new HttpRequestRecord(request, 2L, null, true));
-        records.add(new HttpRequestRecord(request, 3L, null, true));
-        records.add(new HttpRequestRecord(request, 4L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 0L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 1L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 2L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 3L, null, true));
+        records.add(new HttpRequestRecord(httpRequest, 4L, null, true));
 
-        Mockito.when(request.getEntity()).thenReturn(httpEntity);
+        Mockito.when(httpRequest.getEntity()).thenReturn(httpEntity);
         Mockito.when(response.getStatusLine()).thenReturn(statusLine);
         Mockito.when(statusLine.getStatusCode()).thenReturn(500);
         List<HttpSinkResponse> responses = new ArrayList<>();
@@ -118,12 +118,12 @@ public class HttpSinkTest {
         responses.add(new HttpSinkResponse(response, true));
         responses.add(new HttpSinkResponse(response, true));
 
-        when(requestParser.convert(messages)).thenReturn(records);
+        when(request.createRecords(messages)).thenReturn(records);
         List<HttpRequestRecord> validRecords = records.stream().filter(HttpRequestRecord::isValid).collect(Collectors.toList());
         when(httpSinkClient.send(validRecords)).thenReturn(responses);
         when(httpSinkClient.send(records)).thenReturn(responses);
 
-        HttpSink httpSink = new HttpSink(httpSinkClient, requestParser, instrumentation);
+        HttpSink httpSink = new HttpSink(httpSinkClient, request, instrumentation);
         OdpfSinkResponse odpfSinkResponse = httpSink.pushToSink(messages);
         Assert.assertTrue(odpfSinkResponse.hasErrors());
         Assert.assertEquals(3, odpfSinkResponse.getErrors().size());
